Rename infer_call_func to infer_call
