+++ crates/emmylua_code_analysis/src/semantic/infer/infer_call/infer_setmetatable.rs
